Refactor spec into multiple test cases.
The first test case is for versions not supported that are too low for the range of supported versions and the second case is versions that are too high. In the second test case this imitates the behaviour where if a client submits a version that is higher than what the server permits then the negotiation fails but then the client will retry until it can find an acceptable version.