Extract TCP connection initialization into factory
At the risk of adding factories everywhere I have refactored the
creation of the underlying NetworkStream for the TCP connection into a
separate factory method.

This decouples the async creation of the TCP client from the
TcpConnection object.

I have a couple of concerns however, one is that passing only the
NetworkStream may not be enough, especially because we might be causing
a memory leak by creating the TCP client but not disposing of it.

Another reason for this refactoring is also to start to allow
parameterization of the IP address and port used for the TCP connection.