Refactor connection initialization
I've represented initialized connections as concrete classes that can be
passed around and uninitialized as classes that should error if you try
to invoke them in any way.

This separation lets you create a basic connection with opening the
connection and therefore allows the consumer to manage the creation of
connections separately from their use of resources.

I think this makes it more predictable, in that a connection only begins
to consume resources once you actually make a call to open it. Prior to
that it is just another POCO which contains configuration data.