Move protocol header information to a separate class
Previously the protocol header information was mixed in with the
version. I've now extracted this into a separate ProtocolHeader class
which contains a Version object.

I've also extracted the byte value that we use for the protocol version
into a class called ProtocolId so that it can be strongly typed.