Narrow test expectation
Expecting that the array contains the bytes in order is good but these
should be at the beginning of the array as well. We can narrow the test
expectation by one selecting the first four bytes of the array.

We also get to use the nifty new 'range' in C#.